Added code button to menu
