+++ src/screens/home/views/ExposureView.tsx
@@ -1,15 +1,13 @@
 import React, {useCallback} from 'react';
 import {Linking} from 'react-native';
 import {useI18n} from '@shopify/react-i18n';
-import {Text, Box, ButtonMultiline, ButtonSingleLine} from 'components';
-import {useNavigation} from '@react-navigation/native';
+import {Text, Box, ButtonSingleLine} from 'components';
 import {useStorage} from 'services/StorageService';
 
 import {BaseHomeView} from '../components/BaseHomeView';
 
 export const ExposureView = () => {
   const {region} = useStorage();
-  const navigation = useNavigation();
   const [i18n] = useI18n();
   const onActionGuidance = useCallback(() => {
     let url = i18n.translate(`RegionalGuidanceURL.CA`);
@@ -21,8 +19,6 @@ export const ExposureView = () => {
     }
     Linking.openURL(url).catch(err => console.error('An error occurred', err));
   }, [i18n, region]);
-
-  const onDiagnosed = useCallback(() => navigation.navigate('DataSharing'), [navigation]);
 
   return (
     <BaseHomeView iconName="hand-caution">
@@ -45,15 +41,6 @@ export const ExposureView = () => {
           variant="bigFlatPurple"
           externalLink
           onPress={onActionGuidance}
-        />
-      </Box>
-      <Box alignSelf="stretch" marginTop="s" marginBottom="xl">
-        <ButtonMultiline
-          text={i18n.translate('Home.ExposureDetected.DiagnosedBtnText1')}
-          text1={i18n.translate('Home.ExposureDetected.DiagnosedBtnText2')}
-          variant="bigFlat"
-          internalLink
-          onPress={onDiagnosed}
         />
       </Box>
     </BaseHomeView>
